페이지 상대경로 context 이용하여 수정, 마이페이지 성별 표시 수정, 회원가입 후 로그인 페이지 이동 오류 수정

+++ s20230404/src/main/java/com/travelAlone/s20230404/dao/km/MypageDaoImpl.java
@@ -135,7 +135,7 @@ public class MypageDaoImpl implements MypageDao{
 
     @Override
     public int kmMypageFavoritesCountRes(Long id) {
-        return session.selectOne("kmMypageFavoritesCountRes",(long)id);
+        return session.selectOne("kmMypageFavoritesCountRes",id);
     }
 
     @Override
